+++ pkg/rpcservice/version_test.go
@@ -2,8 +2,6 @@ package rpcservice
 
 import (
 	"context"
-	"log"
-	"net"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -11,35 +9,7 @@ import (
 	"github.com/stripe/stripe-cli/rpc"
 
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
-	"google.golang.org/grpc/test/bufconn"
 )
-
-const bufSize = 1024 * 1024
-
-var lis *bufconn.Listener
-
-func init() {
-	lis = bufconn.Listen(bufSize)
-	srv := New(&Config{})
-
-	rpc.RegisterStripeCLIServer(srv.grpcServer, srv)
-
-	go func() {
-		if err := srv.grpcServer.Serve(lis); err != nil {
-			log.Fatalf("Server exited with error: %v", err)
-		}
-	}()
-}
-
-func bufDialer(context.Context, string) (net.Conn, error) {
-	return lis.Dial()
-}
-
-func withAuth(ctx context.Context) context.Context {
-	md := metadata.New(map[string]string{requiredHeader: "1"})
-	return metadata.NewOutgoingContext(ctx, md)
-}
 
 func TestVersionReturnsCLIVersion(t *testing.T) {
 	ctx := withAuth(context.Background())
